+++ src/const.js
@@ -1,26 +1,1 @@
-// const DESCRIPTIONS = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.';
-// const CITIES = [];
-// const Price = {min: 1, max: 1000};
-// const Duration = {HOUR: 5, DAY: 5, MIN: 59};
-// export const WAYPOINT_TYPES = [ 'taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant' ];
-// export const TITLE_OFFERS = [
-//   'Upgrade to a business class',
-//   'Order Uber',
-//   'Add luggage',
-//   'Switch to comfort class',
-//   'Rent a car',
-//   'Add breakfast',
-//   'Lunch in city',
-//   'Travel by train',
-// ];
-
-// export const POINT_EMPTY = {
-//   id: 0,
-//   basePrice: 0,
-//   dateFrom: null,
-//   dateTo: null,
-//   destination: null,
-//   isFavorite: false,
-//   offers: [],
-//   type:'flight'
-// };
+export const TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
