+++ src/model/points-model.js
@@ -1,5 +1,3 @@
-import { getRandomArrayElement } from '../utils';
-
 export default class PointsModel {
   constructor(points) {
     this.points = points;
@@ -7,10 +5,6 @@ export default class PointsModel {
 
   get() {
     return this.points;
-  }
-
-  getPoint() {
-    return getRandomArrayElement(this.points);
   }
 
 }
