+++ src/utils.js
@@ -50,9 +50,6 @@ export function getRandomArrayElement(items) {
   return items[Math.floor(Math.random() * items.length)];
 }
 
-export const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
+export function capitalize(string) {
+  return `${string[0].toUpperCase()}${string.slice(1)}`;
+}
