+++ src/view/form/editing-form-template.js
@@ -1,0 +1,113 @@
+import{formatSrtingToDateTime, capitalize } from '../../utils.js';
+import { TYPES } from '../../const.js';
+
+function showType(types, type) {
+  return types.map((item) => (` <div class="event__type-item">
+  <input ${type === item ? 'checked' : ''} id="event-type-${item}-1" class="event__type-input  visually-hidden" type="radio" 
+  name="event-type" value="${item}" >
+  <label class="event__type-label  event__type-label--${item}" for="event-type-${item}-1">${capitalize(item)}</label>
+</div>`)).join('');
+}
+
+function showOffers(pointOffers, offers) {
+  return pointOffers.map((item) => (`<div class="event__offer-selector">
+  <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-1" type="checkbox" name="event-offer-seats" 
+  ${offers.find((elem) => elem === item.id) ? 'checked' : ''}>
+  <label class="event__offer-label" for="event-offer-seats-1">
+    <span class="event__offer-title">${item.title}</span>
+    &plus;&euro;&nbsp;
+    <span class="event__offer-price">${item.price}</span>
+  </label>
+</div>`)).join('');
+}
+
+function showPhotos(photos){
+  return photos.map((item) => (`<img class="event__photo" src=${item.src} alt="${item.description}">`)).join('');
+}
+
+function destinationList(items) {
+  return items.map((item) => `<option value="${item.name}"></option>`).join('');
+}
+
+export function createFormTemplate(point, pointDestination, pointOffers, arrayDestinationsModel){
+  const {basePrice, type, dateFrom, dateTo, offers} = point;
+  const {description, pictures} = pointDestination;
+  return `<form class="event event--edit" action="#" method="post">
+ <header class="event__header">
+   <div class="event__type-wrapper">
+     <label class="event__type  event__type-btn"
+     for="event-type-toggle-1">
+       <span class="visually-hidden">Choose event type</span>
+       <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
+     </label>
+     <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
+
+     <div class="event__type-list">
+       <fieldset class="event__type-group">
+         <legend class="visually-hidden">Event type</legend>
+
+          ${showType(TYPES, type)}
+
+       </fieldset>
+     </div>
+   </div>
+
+   <div class="event__field-group  event__field-group--destination">
+     <label class="event__label  event__type-output" for="event-destination-1">${type}</label>
+     <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" 
+     placeholder="Chamonix" value="" list="destination-list-1">
+     <datalist id="destination-list-1">
+
+       ${destinationList(arrayDestinationsModel)}
+       
+     </datalist>
+   </div>
+
+   <div class="event__field-group  event__field-group--time">
+     <label class="visually-hidden" for="event-start-time-1">From</label>
+     <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" 
+     value="${formatSrtingToDateTime(dateFrom)}">
+     &mdash;
+     <label class="visually-hidden" for="event-end-time-1">To</label>
+     <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" 
+     value="${formatSrtingToDateTime(dateTo)}">
+   </div>
+
+   <div class="event__field-group  event__field-group--price">
+     <label class="event__label" for="event-price-1">
+       <span class="visually-hidden">Price</span>
+       &euro;
+     </label>
+     <input class="event__input  event__input--${basePrice}" id="event-price-1" type="text" name="event-price" value="${basePrice}">
+   </div>
+
+   <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
+   <button class="event__reset-btn" type="reset">Cancel</button>
+ </header>
+ 
+ <section class="event__details">
+
+ <section class="event__section  event__section--offers">
+ <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+
+ <div class="event__available-offers">
+ ${showOffers(pointOffers, offers)}
+ </div>
+</section>
+
+ <section class="event__section  event__section--destination">
+   <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+   <p class="event__destination-description">${description}</p>
+
+   <div class="event__photos-container">
+     <div class="event__photos-tape">
+       ${showPhotos(pictures)}
+     </div>
+   </div>
+
+ </section>
+</section>
+
+</form>`;
+}
+
