+++ src/view/points-container-view.js
